Extract partitioned array from core into new library
