+++ source/framework/CMakeLists.txt
@@ -3,6 +3,9 @@ list(APPEND doxygen_input ${CMAKE_CURRENT_SOURCE_DIR}/doc)
 add_subdirectory(core)
 list(APPEND doxygen_input ${CMAKE_CURRENT_SOURCE_DIR}/core)
 
+add_subdirectory(partitioned_array)
+list(APPEND doxygen_input ${CMAKE_CURRENT_SOURCE_DIR}/partitioned_array)
+
 add_subdirectory(algorithm)
 list(APPEND doxygen_input ${CMAKE_CURRENT_SOURCE_DIR}/algorithm)
