+++ source/framework/algorithm/include/lue/framework/algorithm/array_like.hpp
@@ -1,9 +1,9 @@
 #pragma once
-#include "lue/framework/core/annotate.hpp"
-#include "lue/framework/core/component/partitioned_array.hpp"
 #include "lue/framework/algorithm/policy/all_values_within_domain.hpp"
 #include "lue/framework/algorithm/policy/default_policies.hpp"
 #include "lue/framework/algorithm/policy/default_value_policies.hpp"
+#include "lue/framework/partitioned_array.hpp"
+#include "lue/framework/core/annotate.hpp"
 
 
 namespace lue {
