+++ source/framework/algorithm/include/lue/framework/algorithm/detail/halo_partition.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "lue/framework/algorithm/flow_direction.hpp"
-#include "lue/framework/core/component/partitioned_array.hpp"
+#include "lue/framework/partitioned_array.hpp"
 
 
 namespace lue::detail {
