+++ source/framework/algorithm/include/lue/framework/algorithm/detail/partition.hpp
@@ -1,6 +1,6 @@
 #pragma once
+#include "lue/framework/partitioned_array.hpp"
 #include "lue/framework/core/annotate.hpp"
-#include "lue/framework/core/component/partitioned_array.hpp"
 // #include "lue/framework/core/type_traits.hpp"
