+++ source/framework/algorithm/include/lue/framework/algorithm/fill.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "lue/framework/core/component/partitioned_array.hpp"
+#include "lue/framework/partitioned_array.hpp"
 
 
 namespace lue {
