+++ source/framework/algorithm/include/lue/framework/algorithm/mesh_grid.hpp
@@ -1,6 +1,6 @@
 #pragma once
+#include "lue/framework/partitioned_array.hpp"
 #include "lue/framework/core/annotate.hpp"
-#include "lue/framework/core/component/partitioned_array.hpp"
 
 
 namespace lue {
