+++ source/framework/algorithm/include/lue/framework/algorithm/range.hpp
@@ -1,6 +1,6 @@
 #pragma once
+#include "lue/framework/partitioned_array.hpp"
 #include "lue/framework/core/annotate.hpp"
-#include "lue/framework/core/component/partitioned_array.hpp"
 #include <hpx/include/lcos.hpp>
