+++ source/framework/algorithm/include/lue/framework/algorithm/unary_aggregate_operation.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "lue/framework/algorithm/functor_traits.hpp"
+#include "lue/framework/partitioned_array.hpp"
 #include "lue/framework/core/annotate.hpp"
-#include "lue/framework/core/component/partitioned_array.hpp"
 
 
 namespace lue {
