+++ source/framework/core/test/CMakeLists.txt
@@ -17,13 +17,3 @@ add_unit_tests(
         lue::framework_core
         lue::framework_test
 )
-
-add_hpx_unit_tests(
-    SCOPE lue_framework_core
-    NAMES
-        array_partition_test.cpp
-        partitioned_array_test.cpp
-    LINK_LIBRARIES
-        lue::framework_partitioned_array
-        lue::framework_test
-)
