+++ source/framework/io/include/lue/framework/io/util.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "lue/framework/core/component/partitioned_array.hpp"
+#include "lue/framework/partitioned_array.hpp"
 #include "lue/hdf5/hyperslab.hpp"
 #include <map>
