+++ source/framework/io/src/read.cpp
@@ -1,5 +1,5 @@
 #include "lue/framework/io/write.hpp"
-#include "lue/framework/core/component/partitioned_array.hpp"
+#include "lue/framework/partitioned_array.hpp"
 
 
 namespace lue {
