+++ source/framework/partitioned_array/test/CMakeLists.txt
@@ -1,0 +1,9 @@
+add_hpx_unit_tests(
+    SCOPE lue_framework_core
+    NAMES
+        array_partition_test.cpp
+        partitioned_array_test.cpp
+    LINK_LIBRARIES
+        lue::framework_partitioned_array
+        lue::framework_test
+)
